unstable checkpoint, remodeling componenets, seeds, and settings structure
