+++ react-app/src/components/Settings_Page/Settings_Page.js
@@ -2,9 +2,9 @@ import { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import { resetProfileTheme } from '../../store/session';
 import { readUserSettings } from '../../store/settings_store';
-import { EditorForm } from './Editor_Form';
-import { SetterForm1 } from './Setter_Form_1';
-import { SetterForm2 } from './Setter_Form_2';
+import { EditThemeForm } from './Edit_Theme_Form';
+import { ProfileMediaForm } from './Profile_Media_Form';
+import { CreateThemeForm } from './Create_Theme_Form';
 import '../Main.css';
 import './Settings_Page.css';
 
@@ -28,14 +28,19 @@ export const SettingsPage = ({ style }) => {
                 className='settings-actions-container' 
                 style={{ border: `3px solid ${style.accent_3}`, backgroundColor: style.background_color }}
             >
-                 <h1 style={{ borderBottom: `2px solid ${style.accent_1}` }}>Profile</h1>
-                <SetterForm1 style={style} />
+                <h1 style={{ borderBottom: `2px solid ${style.accent_1}` }}>Profile</h1>
+                <ProfileMediaForm style={style} />
                 <button data-setting-id='null' type='button' onClick={(e) => resetProfileHandler(e, 'reset_theme')}>Set Theme to Default</button>
             </div>
-                <SetterForm2 style={style} />
+            <div 
+                className='create-theme-container' 
+                style={{ border: `3px solid ${style.accent_3}` }}
+            >
+                <CreateThemeForm style={style} />
+            </div>
             </div>
             <div className='themes-container'>
-            <EditorForm style={style} />
+            <EditThemeForm style={style} />
             </div>
         </div>
     )
